Organize server module into packages, validate player nick
