+++ server/src/main/kotlin/com/cyberbot/bomberman/server/Server.kt
@@ -1,5 +1,6 @@
 package com.cyberbot.bomberman.server
 
+import com.cyberbot.bomberman.server.control.ServerService
 import org.apache.logging.log4j.kotlin.Logging
